enable dependabot for GitHub actions (#37)
* enable dependabot for GitHub actions

This allows us to get updates for GitHub actions automatically. I have used this for my own packages as well as our [Trixi.jl framework](https://github.com/trixi-framework). After merging this, we should also enable other Dependabot actions in "Settings -> Code security and analysis -> Dependabot alerts" and "... -> Dependabot security updates".

After merging this, Dependabot will prepare PRs with updates of the versions of GitHub actions we use such as.

I think it would be nice to have this in all of SciML. To get the discussion started, I picked this repo since the CI tests are not very expensive.

* format